+++ lib/time_tracker.dart
@@ -1,3 +1,3 @@
-export 'src/base/time_tracker.dart';
+export 'src/mixin/time_tracker.dart';
 export 'src/enum/time_status.dart';
-export 'src/extensions/equal.dart';
+export 'src/extension/equal.dart';