+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 1.1.0
-- Requires Dart ^3.0.0
+- Require Dart ^3.0.0
+- `TimeTracker` is now a mixin.
+- Updated docs.
 
 ## 1.0.1